+++ src/main/java/com/b1n_ry/yigd/config/OnDestroyedDrop.java
@@ -1,0 +1,7 @@
+package com.b1n_ry.yigd.config;
+
+public enum OnDestroyedDrop {
+    NONE,
+    GRAVE_BLOCK,
+    ITEM_CONTENTS
+}